+++ lib/benchmark/bigo.rb
@@ -5,6 +5,7 @@ require 'csv'
 require 'benchmark/ips'
 require 'benchmark/bigo/report'
 require 'benchmark/bigo/chart'
+require 'benchmark/bigo/termplot'
 require 'benchmark/bigo/job'
 
 module Benchmark